chore(racingcar): restructure project directory
