+++ src/main/kotlin/racingcar/Application.kt
@@ -1,6 +1,9 @@
 package racingcar
 
+import racingcar.console.Input
+import racingcar.domain.Game
+
 fun main() {
-    val settings = Input.readCarNamesToRound()
+    val settings = Input.readSettings()
     Game.play(settings)
 }
