+++ src/test/kotlin/racingcar/console/RoundInputTest.kt
@@ -1,9 +1,9 @@
-package racingcar
+package racingcar.console
 
 import camp.nextstep.edu.missionutils.test.Assertions.assertSimpleTest
 import org.assertj.core.api.Assertions.assertThat
-import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
+import org.junit.jupiter.api.Test
 import support.IOTest
 
 class RoundInputTest : IOTest() {
